refactor: extract fastapi and streamlit entrypoints in dedicated app layer.
