+++ vigil/app/fastapi/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI HTTP entrypoint."""
