+++ vigil/app/streamlit/__init__.py
@@ -1,0 +1,1 @@
+"""Streamlit UI entrypoint."""
