+++ vigil/monitoring/__init__.py
@@ -1,0 +1,5 @@
+"""Monitoring bounded context.
+
+Tracks the progress of video analysis jobs by reacting to domain events. Exposes
+analysis status (queued, in-progress, complete) to the outside world.
+"""
