+++ vigil/shared_kernel/__init__.py
@@ -1,0 +1,1 @@
+"""Shared kernel, cross-context building blocks."""
