+++ vigil/video_analysis/__init__.py
@@ -1,1 +1,4 @@
-"""Video analysis context."""
+"""Video analysis bounded context.
+
+Handles video ingestion, object detection, and multi-object tracking.
+"""
